Homework 4

+++ training-diary/src/main/java/com/ylab/intensive/model/dto/StatisticsDto.java
@@ -19,6 +19,6 @@ public class StatisticsDto {
     /**
      * The sum of calories burned during workouts.
      */
-    private int sumCalorie;
+    private float sumCalorie;
 }
